+++ test/testing_test.rb
@@ -71,7 +71,7 @@ class TestingTest < Minitest::Spec
     test_case = test.new(:test_0003_anonymous)
     failures = test_case.()
 
-    failures[0].message.must_equal %{assert_call expected not_right terminus, not #<Trailblazer::Activity::End semantic=:failure>. Use assert_call(activity, terminus: :failure).
+    assert_equal failures[0].message, %{assert_call expected not_right terminus, not #<Trailblazer::Activity::End semantic=:failure>. Use assert_call(activity, terminus: :failure).
 Expected: :not_right
   Actual: :failure}
 
@@ -80,7 +80,7 @@ Expected: :not_right
     test_case = test.new(:test_0004_anonymous)
     failures = test_case.()
 
-    failures[0].message.must_equal %{--- expected
+    assert_equal failures[0].message, %{--- expected
 +++ actual
 @@ -1,3 +1,3 @@
  # encoding: US-ASCII
@@ -132,7 +132,7 @@ Expected: :not_right
     test_case = test.new(:test_0002_anonymous)
     failures = test_case.()
 
-    failures[0].message.must_equal %{--- expected
+    assert_equal failures[0].message, %{--- expected
 +++ actual
 @@ -1,3 +1,3 @@
  # encoding: US-ASCII